Refactor ambient sound folder structure
