+++ app/views/layouts/application.html.erb
@@ -65,6 +65,6 @@
     </main>
     <%= render partial: "layouts/issue_detail", locals: issue_detail_context %>
 
-    <%= react_component("FocusApp") %>
+    <%= react_component("FocusSpace/FocusSpace") %>
   </body>
 </html>
