+++ frontend/components/FocusSpace/AmbientSounds/index.js
@@ -1,0 +1,2 @@
+export * from './AmbientSounds'
+export { default } from './AmbientSounds'
